Use PatternFly componentized Login to support custom brang img
