+++ frontend/hub/main/HubLogin.tsx
@@ -1,7 +1,7 @@
 import { Page } from '@patternfly/react-core';
 import { mutate } from 'swr';
 import { LoadingState } from '../../../framework/components/LoadingState';
-import { Login } from '../../common/Login';
+import { AnsibleLogin } from '../../common/AnsibleLogin';
 import { useHubActiveUser } from '../../hub/common/useHubActiveUser';
 import { hubAPI } from '../common/api/formatPath';
 import { HubContextProvider } from '../common/useHubContext';
@@ -19,15 +19,14 @@ export function HubLogin(props: { children: React.ReactNode }) {
 
   if (!activeHubUser) {
     return (
-      <Login
-        apiUrl={hubAPI`/_ui/v1/auth/login/`}
+      <AnsibleLogin
+        loginApiUrl={hubAPI`/_ui/v1/auth/login/`}
         onSuccess={() => {
           refreshActiveHubUser?.();
           void mutate(() => true);
         }}
-        icon="/static/media/galaxy-logo.svg"
-        brand={process.env.BRAND}
-        product={process.env.PRODUCT}
+        brandImg="/static/media/galaxy-logo.svg"
+        brandImgAlt={process.env.PRODUCT}
       />
     );
   }
